Improved validate_all_cpan test
We no longer shell out to another perl to work out if a module name is
valid, we just try to declare a package with that name. This has the
side-effect of creating lots of empty packages, but meh.

The following invalid modules were picked up in the last test run:
 * Date:Jalali
 * HTTP::Server:::Brick
 * Net:::DDDS
 * Net::SMS::Clickatell:SOAP
 * Time:Compare
 * WWW::Mechanize:Frames
 * XAO::DO::Web:IdentifyAgent